refactor: 온보딩 플로우 로직, 공통 셀렉트 필드 분리 (#12)
